feat(database): Add `UserEntity`
