+++ apps/api/src/database/database.module.ts
@@ -1,6 +1,7 @@
 import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
 
+import { User } from '@szklarnia-pwr/database';
 import { ConfigModule, ConfigService, NodeEnv } from '../config';
 
 @Module({
@@ -12,7 +13,7 @@ import { ConfigModule, ConfigService, NodeEnv } from '../config';
                 ...configService.databaseCredentials(),
                 // logging: configService.NODE_ENV === NodeEnv.DEVELOPMENT,
                 synchronize: configService.NODE_ENV === NodeEnv.DEVELOPMENT,
-                entities: [],
+                entities: [User],
             }),
         }),
     ],
