+++ packages/database/src/entities/index.ts
@@ -1,1 +1,1 @@
-export * from './user.entity';
+export * from './user';
