+++ packages/database/src/entities/user/index.ts
@@ -1,0 +1,2 @@
+export * from './user-role.enum';
+export * from './user.entity';
